[F] Allow Form components outside Form.Form
Prior to this change, all of our custom form components would need to be
inside of a Form container to function properly. That container is responsible
for passing the set action down to the form components, which is used by
the Set component to update the session model. With this change, the Set
component will behave in a pass-through way, just rendering the form
element without hooking into its onChange event, if it hasn’t been passed
a set action to work with.