[C] Refactor filtering concern
Prior to this commit, logic around filtering and pagination was spread
between the model, the filterable, and the pagination concerns. With
this revision, we abandon the pagination concern and move all logic
related to filtering and querying into the Filterable concern. The concern
will look for named scopes beginning with `by_` or `with_` on the model
and map filter params to those scopes. If keyword or typeahead is
present, the search will be further delegated to Elastic search. Because
pagination is tightly coupled with filtering, the pagination-related code has
also been moved into the filterable concern.