[C] Refactor request meta keys
This commit normalizes all API calls that go through the store so that
each call uses a constant defined in the entity actions `requests`
object. This refactoring sets the stage for being able to handle
notification messages based on the request key rather than handling
these in every component that makes an API call.