[B] Fix hash links in reader
This commit makes a significant change to how the reader fetches data.
With the reader we don’t want to trigger fetchData when hashes in the
URL change. This may not be a long-term change. The bigger problem is
a question of how we pass a new location to components rendered by the
router without triggering a new fetchData call. This doesn’t address that
problem per se, but does provide a work around for the time being.

[Delivers #138746857]