[F] Improve RequireRole interface
With this commit, the RequireRole component accepts a role of “none”
which will cause it to show children only for unauthenticated users. The
component will also accept a roleMatchBehavior prop which can be hide
or show. If the value is hide, it will not render children when the role
matches, and them when it does not. Default behavior is to show children
when the role matches.