[F] Improve global image handling
Note to @naomi and @max.ono… I’ve refactored all of the attachment
handling in the Manifold API and will be pushing it to development
shortly. Prior to this commit, there were attributes like
“attachmentUrl” or “avatarUrl” on models, which mostly pointed to the
high res original file. There was a lot of inconsistency as to how
these files were resized and what variants were available for each, and
the attachment config on the models was kind of all over the place. To
clarify this, I’ve written an attachment concern which will setup the
attachments the same way in the API. It will detect if any given
attachment is an image, and if it is, it will produce the same set of
variations. To add an attachment field to a model, just do this:

```
# Include the model concern
include Attachments
# Setup the field
manifold_has_attached_file :avatar, :image
```

The concern will add the following methods to the model, which can be
used in the API or may be exposed to the client during serialization.
```
avatar_url
avatar_extension
avatar_styles
avatar_is_image?
can_process_avatar?
```
For models with attachments, the client will see a consistent response
that looks like this:
```
"heroStyles": {
"small": "http://some/path",
"smallSquare": ""http://some/path",
"smallLandscape": ""http://some/path",
"smallPortrait": ""http://some/path",
"medium": ""http://some/path",
"mediumSquare": ""http://some/path",
"mediumLandscape": ""http://some/path",
"mediumPortrait": ""http://some/path",
"largeLandscape": "http://some/path",
"original": "http://some/path"
}
```
I’ve gone through every place in the client where an image is included
and adjusted it so it’s consistent and uses this new approach.