[F] Whitelist instead of blacklist tag attributes
Prior to this commit, the HTML validator, which is used during
text ingestion, blacklisted certain tag attributes. We can’t
possibly blacklist all the incorrect attributes that could exist
in an EPUB, so we’re switching to a whitelist approach, in which
we specify exactly which attributes we’ll accept.