[F] Add request result to existing store response
This commit exposes an “adds” option on the EntityStore request
options object. The value of adds can be the meta of a previous
response. If the request resolves to an entity, it will be added
to the end of the previous response.

This option is useful for cases where a user is creating an
entity in the client, and we don't want to retry a request to
add the new entity to it. For example, if a user is adding a
comment to a comment thread, we shouldn't have to request the
thread all over again to have the comment enter the store.