[F] Build out annotation and commenting UI
With this commit, users are able to create full annotations, including
body copy. It lays the foundation for comment threads in the client
application, altough this work is not yet complete. This revision also
includes improvements to drawer scroll behavior, and implements the
visibility button in the reader header. Annotatable is refactored
somewhat, as is the drawer interface.