[B] Improve ENTITY_STORE_ADD reducer behavior
In cases where an ENTITY_STORE_ADD action is adding an entity to a
response that previously contained no results, a subsequent
ENTITY_STORE_REMOVE action would not actually remove the entity,
because the response is lacking a "type" property. When responses
are reduced, the reducer sets the response entity type based on the
type value of the first entity in the response, with the assumption
that responses are always homogeneous. This revision adjusts the
reducer so that when we add an entity to an empty response collection,
we're sure to update the response entity type value.