[B] Store pending loader removals
Sometimes, especially in local development environments, a STOP_LOADING
action can be reduced before the corresponding START_LOADING action is
reduced. This leads to a START_LOADING that never stops. This revision
stores STOP_LOADING actions in a pending queue in the store, which
blocks the corresopnding START_LOADING action from being added to the
active loaders

[Delivers #142145903]