[F] Form components can be used without name prop
This commit changes the setter HOC so that if it is not given a name
prop, it essentially treats the wrapped component as a passthrough.
This allows developers to use things like the switch component outside
of a Form that's tied to an editing session. Setting the name property
on a Form component will, however, activate the setter HOC.