[C] Form has predictable session name
With this change, the Form container will name its editing sessions to
match the name prop that's passed into the form. This is useful because
it allows a parent component that's wrapping the form to use that name
to access the result of a create or update API call in the entity
store.