[F] Backend editors can ingest texts
This is a large commit that includes a number of underlying features.
It introduces an action cable websocket server to Manifold, and adds
the first channel to the application. The Ingestion interfaces are
tweaked a bit, and a new Ingestion model is introduced to track the
state of the ingestion process. New backend interfaces for creating
ingestions and processing them are included as well.