[F] Upgrade react-router to v4; refactor fetchData
This commit upgrades react-router to v4 and fully implements global
link handlers, which are located close to the routes themselves.
Furthermore, it moves container data fetching into a higher order
component, which is easier to understand, and rewrites much of the
server-side-rendering approach. This is an extensive refactoring
of the routing and rendering methods, and as such touches many parts
of the client-side application. This commit also begins introducing
human-readable URLs for #275, although there is additional work to be
done on that front.

Fixes #310: 	Typekit fonts should load (if configured) when SSR is
				absent
Fixes #309: 	Use custom dialog box for confirmation on leaving forms
		   		without saving first.
Fixes #304: 	Component FetchData methods should not be triggered by
				URL hash changes
Fixes #276: 	Update react-router from 2.x to 4.x
Fixes #264: 	Improve scroll behavior on route change
Starts #275: 	Implement semantic URLs