[F] Setter HOC passes setOther prop to child
With this commit, the Setter HOC that sits between the Form container
and child form input components will now pass as `setOther` prop to
the wrapped component in addition to the `set`, `onChange`, `value`, &
`initialValue` props that were already passed. The `setOther` prop
accepts value and name (in the form of `attributes[name]`), and will
trigger updates to the editing session dirty model. The intended use
is for complex form components that need to manage more than one
attribute on the model being edited, such as the project avatar
builder component.