[B] ResourceViewer does not display resources without SSR
This revision addresses a couple issues. Prior to this change, no
resource thumbnails were visible in the reader margins unless they
were present in the SSR load.

The cause for this was a bit involved. The ResourceViewer needs to
interact with the DOM to measure the height of the section body and
to position resource thumbnails in the correct spot in the margin.
Prior to this change, the Text container was passing a ref down
through its children. This was a less-than-ideal solution because in
many cases it was passing the ref down before the ref existed,
because the component had not yet mounted.

This commit passes a selector instead of a ref down to the
ResourceList. It also introduces a timer that checks the height of
the body container and repositions the resources if the height
changes. This is necessary for images embedded in text sections.
When those images load, the resources have to move to account for
the changed height of the section.

Closes #252