[F] Improve reader citation generator
With this commit, citation generation is moved from the client to the
server. Citations are generated using the Ruby CiteProc gem. If a model
can be cited, it should implement a citation_parts method that is used
to generate the citation. We now have a "citable" model concern that
simplifies this process.

Closes #269