[B] Adjust CSS validation rules for Google Docs
This commit disallows line-height, font-size, and various other rules
to be applied to paragraph tags. These declarations were previously
being stripped from headers, and now we’re extending it to paragraphs
as well. This is motivated by the poor quality of texts ingested from
google, which are full of inline, hard-coded styles that lead to a
poor reading experience in Manifold.

Fixes #368