[B] Fix issue with resource marker not updating on delete
This commit takes advantage of the recent addition to the entity store
middleware and reducer that allows a dispatched API call (usually for
create requests) to add the created resource to the results of a
previous request. For example, in the case of annotations, we don't
want to have to re-fetch the collection of all annotations when a new
annotation is created. With the "adds" request option, the resulting
annotation from the create request is added to the results of the
rAnnotation request.

This commit also removes older, hacky code in the entity store reducer
that was there specifically to handle this annotations use case.

[Fixes #373]