[C] Exclude generated chunks from node app servers
As we start to incorporate more third party libraries for specific
features, we need a mechanism for spliting library code out into
separate chunks. A subsequent commit will use webpack 1's
require.ensure feature to split the code for the Ace editor into a
separate chunk. All code that is split into a separate chunk will be
loaded asynchronously via Webpack's jsonp approach. This code will
not be necessary in the server-side app, however, and should not
be included in the server-side build. Prior to this commit, the
universal-webpack module we're using was collapsing all chunks into
a single file on the serve. The change in this commit monkey patches
our webpack config so that the chunks are output to separate files,
which are ignored by the server-side application.

This whole approach will likey change, of course, when we upgrade to
Webpack 2. Yay Javascript churn.