[F] Add withDispatch HOC
This higher order component can be used to inject redux's dispatch
function into components that are otherwise unconnected to the store.
While we understand the value of keeping containers and components
separate, we've come to think that turning a component into a container
just so it can dispatch an action is often overkill.