[B] Changing position updates model property
See https://github.com/swanandp/acts_as_list/issues/23 for context. We
were seeing an issue in which moving a model up in the list would not
update the models position property if there were two items in the list.
For now, we're reloading the model after changing the position to ensure
an accurate position property.