[F] Refactor and improve ingestion; add specs
This commit moves all file-level access to ingested sources out of the
individual strategies and into the ingestion object. This change
simplifies code across all the strategies, because it allows each
strategy to access the ingestion source via relative package paths.
Prior to this change, each strategy was responsible for unzipping
packages, translating relative paths to absolute paths, etc. This
change also makes ingestion safer, as it gives us a central place to
validate relative paths to ensure that the ingestion doesn't escape
the extracted or copied package directory. All ingestion strategies
are now compatible with an extractable (zip) file, or with a source
directory.

This commit also adds integration specs for each strategy and improved
sample ingestion sources.

Finally, it removes Word ingestion, as there was overlap between HTML
and Word ingestion. Along these lines, the Google doc ingestion, which
pulls the document in as HTML, now delegates almost all functionality
to the HTML strategy. As such, it's an example of how we can write
strategies that rely on other strategies.

Resolves #399
Resolves #398
Resolves #397