[F] Remove absolute font sizes from ingested styles
We don’t have a reliable way to handle absolute font sizes, and trying
to map absolute units to relative units requires more effort than we
want to invest at this time. This commit updates CSS validation to
remove any font-size properties that are measured in the absolute units
of pt or px. Furthermore, it strips all line-height declarations. It is
our thinking that controlling these aspects of the reading experience
outweighs the downsides.

To accomplish this, we add a new property condition that's based on
declaratino value to the CSS validator configuration.

Fixes #388