[C] Add withFormSession HOC
Wrap a component with this HOC and pass it the name of a form. Doing
so will pass a form prop to the wrapped component with access to the
editing session with the dirty model, a changed flag, and a
getModelValue method. One use for this is for handling conditional
logic in a form based on the currentn state of the form.