[F] Admins can configure and test how email is sent
This revision exposes a new backend “email” view. With this view,
admin users can choose either SMTP or Sendmail for email delivery.
They can also set from and reply-to details for emails, as well as
the closing text for every email. This commit adds an email test
mechanism to ensure that mail is working as expected, and adds a
basic CSS/HTML layout. The deliver_method for the Rails application
is switched over to a dynamic method provided by Manifold. This
dynamic method looks at settings and delegates to underyling SMTP
and Sendmail services as appropriate. Finally, we're adding the
premailer gem, which moves external styles into emails when they're
sent, which improves the developer experience.

Fixes #440