[F] Support epub:type attributes in EPUB3 sources
Because HTML5 does not really support XML namespaces, we'll transform any epub:type attributes to data-epub-type. We need to also adjust selectors in stylesheets when they are parsed and validated.

This is motivated by the realization that the standard ebooks project relies heavily on this attribute as a selector for styling content in texts.

Fixes #460