[F] Allow meta attributes on create and update
The JSON API spec allows us to include a meta object on the payload
when API clients are creating or updating resources. Per the spec:
"Where specified, a meta member can be used to include non-standard
meta-information. The value of each meta member MUST be an object
(a “meta object”)." This revision lays the groundwork for using this
in subsequent commits. For example, we might want to influence the
behavior of what kind of email a controller sends in response to a
request. That can be coded in the request meta. This commit
anticipates using this feature for distinguishing, at the controller
level, between models that were self-created and those that were
created by an admin.