[F] Refactor and improve account emails
This revision updates the copy in the welcome and password reset
messages. It adjusts the welcome mailer so that it only sends the
user's credentials if the account was created by another (admin)
user. All existing account emails are given the new layout and
styles. Preview mechanisms are incorporated for developing and
testing the emails.

Fixes #474