[C] Refactor client servers
Prior to this commit, in both dev and prod environments, Manifold
would run two servers (web and renderer). The Renderer server was
responsible for the isomorphic rendering of the react app. When
client-side code changed, webpack would build a new renderer JS
bundle and the server would be restarted. This mechanism broke hot
reloading, because we're proxying assets through the node server.
To address this, we implemented the "web" server, which is a simple
express server that can continue to run (and proxy the webpack
dev server) while the renderer was restarting. This approach is,
to some degree, necessary. However, it was used in production as
well, which didn't make sense.

With this new approach, we rename the main node server to "client"
and we call the wrapper around it, used only in development, the
"development" server. The development server is never restarted
due to code changes. If it's unable to proxy to the main client
server, it simply renders the page without the isomorphic render, so
that hot reloading can continue to work. In dev, the development
server hijacks the main client port and bumps the client server to
a fallback port.

In production, we now only build and run the client server, and it
runs on the main port.

Furthermore, this commit revises all services to reduce code
redundancy. Both servers share code for generating the express app,
as well as code for generating the HTTP server object. Both servers
now proxy to static assets paths in the same way, which makes for
more predictible outcomes.

This commit also adds support for correct signal handling. These
serves will now catch INT and USR2 and TERM signals and clean up
after themselves. This graceful shutdown is especially important for
production, where we want to run the servers on a unix socket.