[C] Prevent eslint error when using rest property
It’s common in our code to access all but some of an object's
properites using ES6 rest syntax. For example:

const { foo, bar, ...other } = myObject;

ESLint will typically throw an error if foo and bar are not
subsequently used in code. This change suppresses that error if the
unused vars are siblings of a rest property.