[B] Correct Form.Date handling of null values
This commit makes an API change to the Form component. The Form
component will now pass a submitKey prop to its children. This
prop is initially null. Each time a form is submitted, a new six
character key is generated and passed to children. Children can check
the key value to determine a) if the form has been submitted once, and
b) whether the form has been submitted again.

The date picker component needs this, because in the event that the
user has entered an invalid or partial date and then submitted the
form, the input should be cleared to indicate that no value was saved.
This should only happen on submit, or else it's impossible for the
input to build a valid date, since any incoming null value woulc clear
it.

Fixes #554