[B] Correct browser bootstrapping
With this commit, Manifold will no longer attempt to bootstrap, which
entails loading settings and authenticating a user from a cookie, in
the browser if the bootstrapping has already occurred server-side.
Moreover, this commit refactors the App and Manifold components so
that more of the bootstrapping process is located in the Manifold
component.

Fixes #605