[B] Pass path rather than URL to router on SSR
This change is made on the basis of this issue in React Router:

https://github.com/ReactTraining/react-router/issues/5662.

By passing the entire URL to React Router for the server-side render,
we were running into cases where the route was not being matched as
expected. For example, if the route path was "/projects/:id" and the
url was http://manifold.dev/projects/some-project?foo=bar, the server-
side render would be passed a project param of `some-project?foo=bar`
rather than the expected `some-project`. This change solves that
problem, but I'm concerned that we're no longer passing any query
params from the URL to the server-side render, which may lead to mis-
matches between SSR and CSR.