[F] Remove URL hash when annotation is hidden
If there is an annotation hash present in the URL, the text node that
contains the corresponding annotation will trigger a scroll when it is
mounted or when the hash is added to the URL. This approach has proven
to be more reliable than trying to trigger scroll from higher up in the
reader hierarchy, based on dom queries for elements that may or may not
have been rendered.

That challenge, as is often the case with React's unidirectional data
flow, is ensuring that this scroll event only happens one time. A
representiative problem case is as follows:

1. The user clicks on an annotation in the notes drawer, which adds a
"#annotation-1111-2222-3333-4444" hash to the URL.

2. The reader loads the requested section and renders it.

3. Lower down in the render tree, the corresponding text node is
rendered and scrolls itself into view.

4. Next, the user hides all annotations using the visibility menu, and
scrolls around a bit.

5. When the user makes annotations visible again, the annotation is
mounted, and it sees the hash in the URL and triggers another scroll.
This second scroll is not desirable.

We could always render annotations on the page and just hide them with
CSS, but this would require a lot of extra logic within the text nodes.

A better solution, which is what we do in this revision, is to remove
the annotation hash from the URL when the annotation it references is
not visible on the page (kind of like a redirect). When annotations are
filtered in the Text component, this commit will validate the URL hash
and, if necessary, modify it. In doing so, it passes a noScroll flag in
the new location state so that the reader knows not to scroll.