[B] Upgrade node humps module
89f9e6 introduced a regression. In 89f9e6 and following, params in API
requests were decamelized before being stringified for the frontend API
client. On the server-side, it appears that the params object has been
mutated so it no longer has a 'hasOwnProperty' method. Prior to 2.0.1,
Humps used a hasOwnProperty function check to determine if the argument
is an object. See domchristie/humps#35 for details.