[B] Do not force UTF8 epub source encoding
Prior to this revision, we were forcing UTF8 encoding for all ingested
epubs. This caused problems for an Enuiadi ingestion, and I don't see
the reasoning behind forcing this. In the event that the encoding
specified in an EPUB ingestion source differs from the actual encoding,
we should ask the EPUB author to correct the error.