[WIP] Initial pass for api and proxy
Several assumptions are made most importantly that docker.manifold.lvh points to 127.0.0.1 which
can be configured in the /etc/hosts file. This is a POC setup for using named volumes and will be replaced
once the docker-compose.yml implementation is complete (currently in-progress).

To test this out, run `./docker-setup` from within the manifold project root.