[B] Prevent unnecessary entity editor mutation
Prior to this change, the entity editor used lodash's cloneDeep method
to update the dirty model. This approach was problematic because in
cases where the property on a model was an object or an array, the
property would always be updated, which triggered unnecessary (and
at time problematic) updates in form components. This change mutates
the dirtyModel itself every time, but only mutates the properties when
necessary. The change is supported by tests showing the desired
behavior of this reducer.