[B] On logout, flush all entities from store
Now that entities have abilities attached to them that are based on the
current user, we need to remove entities from the store when users
change. We accomplish this by flushing the models when the user logs
out. This means that some containers, in particular the Frontend
container, needs to re-fetch its data after the store has been flushed.
To accomplish this, the FetchData HOC now exposes a fetchData prop to
the wrapped component. The Wrapped component can call this at any time
to trigger a refetch.