[C] Replace Express with Connect; bundle modules
Before this revision, the server-side node application did not include
external modules in the generated bundle. This made for a smaller
bundle, but it also meant that we had to include hundreds of node
modules in our package distributions. With this approach, we build a
single javascript file for the node app, which will allow packages to
exclude node modules, as well as ensure the integrity of the compiled
client application.

To accomplish this, we needed to replace Express with Connect. Express
doesn't fully support being packed by webpack. Moreover, we were only
relying on Express for its middleware. Connect, which is more light-
weight, is a more appropriate option for Manifold.

Fixes #957