[F] Add alt text to non-text non-img elements
Resolves #1061

This mostly includes updates to interactive elements containing no text but only an icon or other non-text visual element.
This also includes hiding any decorative icon so that a screen reader does not need to "tab over it" as the user to exploring the site (i.e. the user only has to survey those elements critical to understanding the site content)