[C] Refactor project authority abilities
We determined that using `update_[entity]` was not indicative of the
actual abilities being checked.  This refactors the `update` name to
`manage_[entity]` and `create_[entity]`.  For the most part, the logic
for both of these cases is the same, but that will likely not always be true.
This also allows our component checks to be more explicit about what
ability they are actually checking.