[C] Refactor backend redirects
Prior to this commit we included redirect logic in backend containers
using the RedirectIfNoChildRouteMatches component. That component was
confusing and unnecessarily complex. This commit replaces it instead
with a new component called RedirecToFirstMatch. RedirectToFirstMatch
is passed the array of sidebar links, and it will redirect the user to
the first path that is allowed per the user's abilities. This change
reduces the possibility for divergent behavior between the secondary
navigation and redirects to the default child page (often /general).