[B] Overlay de-focus should not trigger close
We recently wrapped our overlays in the FocusTrap component as part of
our accessibility efforts. As part of that implementation, we adjusted
the overlay so that it automatically triggers its close handler when
it loses focus. This is a problem because the close handler generally
will trigger a "back" action in history, or will redirect the user to
the URL from which the overlay was opened. The unintended side effect of
this is that if there is an link in the overlay (as in the case of the
resource notation overlay in the reader), clicking that link will
trigger the close behavior rather than going to the intended link
target.

Fixes #1183