[B] Fix left side notation positioning in ie11
Resolves #1194

These were not positioned or styled correctly due to the use of an unsupported property: scrollY. Since the only uses of this property were on the window (i.e. window.scrollY) I replaced it with window.pageYOffset which has total support across browsers (and in  most cases is just an alias for scrollY).