[WIP] Update global backend type, form, and container styles
This commit makes global CSS changes to colors, type, and containers
in the backend UI. It especially makes changes to forms, switching to a
two-column layout for input groups and rearranging some form elements.
To accomodate design changes, it also revises some React components, mainly
to add new classNames to work with the two-column layout.