[B] Fix contents button behavior regression
Resolves #1271
Resolves #1275

Updates the focus trap behavior to allow the contents button to toggle the ToC when clicked (even if the ToC is open).
Clears focus when user selects a different section.
Updates the focus state styles to be more subtle (ToC button and Menu button in the reader).
Updates the focus trap to allow interactive items outside of it to be clicked.