[B] Drawers always have close buttons
Resolves #1280

Adds a screen reader only close button to the drawer even if there is no visible drawer header or close button.
Includes update to the Drawer component to optionally not initialize focus trap.  This is used on the "notes" ReaderDrawer as that drawer already lives in another focus trap supplied by its parent UIPanel.