[B] Fix focus trap return focus for UIPanels
Resolves #1279

The focus trap expects you to be mounting and unmounting the target element when activating and deactivating focus trap. So, when the focus trap mounts/activates the first time it uses whatever element had focus before it activated as the element to return focus to when it is deactivated.  As such if the element was mounted on page load, in the case of the UIPanels, the "return focus to" element defaulted to the document body, and, when the focus trap is toggled manually, it returns focus to the body instead of the element that opened the panel.  This fix mounts and unmounts the panels instead of hiding with css and uses CSSTransitionGroup, along with some new css animations, to emulate the existing slide and fade transitions used with panels.