[FIXUP] Notification feature fixes
This commit fixes a number of issues I discovered with this branch:

- Insufficient test coverage for the jobs made it very difficult to
have confidence that they were doing what they should without actually
triggering them.

- I found the names of the jobs and the structure of the jobs rather
confusing. It took me some time to unwind what each job was doing, and
why in some cases email was being sent from jobs and in other cases it
was sent from services. I rewrote all the notification jobs and
restructured them so that they were queuing other jobs in a
predictable way. I made the jobs each have a single concern so that
they could be tested. I wrote tests for each job.

- I made a number of revisions to language in the notification
preferences form

- Simplified the AI services for reply, flag, and comment notification
services so that they all return a list of recipients, rather than
doing multiple things. In some cases, logic around who should receive
a message was split between jobs and AI services, which was confusing.

- Reworked the AI service specs in keeping with the refactoring I did.

- The jobs triggered by clockwork were all failing in your
implementation because you were passing a NotificationFrequency
instance to the jobs, which active job is unable to serialize. Always
pass simple variable types to jobs as arguments (eg, a string or
integer ID rather than a model instance) as serializing the arguments
is fragile and can lead to trouble. This led to refactoring the jobs
and the inputs to the notification AI services.

- At least two of the notification mailer previews were throwing
exceptions as is due to errors in the views. I fixed these errors.

- None of the text versions of the notification mailers were working
because you were using a txt extension instead of a text extension. I’m
surprised you couldn’t see this, since you must have tested them at
some point.

- The new URL service had incorrect URLs for projects, which I fixed.

- The text versions of the daily and weekly digests were incomplete.
I pretty much rewrote all the text emails to improve consistency and
legibility.

- Fixed places where instead of pluralizing a string, you added an “s”
in the view; corrected places where verb tense didn’t agree with the
subject.

- Added validation to notification service frequency inputs; added
rdoc comments to the jobs to make it clear what kind of arguments they
accept

- Corrected the weekly digest email which was trying to call Text
methods on Project models, triggering exceptions

- Fixed after_create callbacks in models that should be after_commit.
Using after_save leads to possible race conditions where the job is
triggered before the model is committed to the DB.

- Simplified callback logic in the comment model.

- Unsubscribe token is only assigned to the view in the
NotificationMailer, yet you included the link in all emails via the
layout. The result was a broken unsubscribe link on every email other
than notifications. I corrected this.

- Saving the notification form in the client was confusing because the
user is not scrolled to the top. Fixed it.