[B] API sets client URL from CLIENT_URL or DOMAIN
There’s a little ENV cleanup that needs to happen at some point. Our
Omnibus packages don't ask the user to set the domain, instead asking
for the URL of the site in /etc/manifold/manifold.rb. That setting
flows into the API environment as CLIENT_URL. In our local dev
environments and installations installed by puppet, we're setting the
DOMAIN environment variable, and building the client URL in an
initializer based on that domain and whether or not Manifodl is
configured to use SSL. This change allows the interal client URL in the
API to be set from either CLIENT_URL or DOMAIN, in the interest of
backwards compatibility. This needs to be set correctly for the
email notifications to have correct image pages.

Fixes #1381