[F] Expose concept of “slots” in child routes
This revision add the concept of "slots" to child routes, which is
useful in cases where a component wants to render some child routes
in one spot, and others in another spot. By default, a childRoutes()
call will render all child routes, as in the past. If a route has a
"slot" property, that route will only render if the childRoute options
hash has a "slot" property with a matching value. For example

```
<div>
  {childRoutes(this.props.route, { slot: "drawer", drawer: true })}
  {childRoutes(this.props.route)}
</div>
```
The first childRoutes call will render any child routes with a slot
property of "drawer". The second call will render any child routes
with a null or undefined slot property.