[F] Add TagList component
This adds a new TagList component that renders a string or array
into an input with an item bucket below.  It uses the same styles
as a HasMany, the difference being that there is no fetch and that
the value is set on the model.  This means that changes to this
input mark the model as dirty and are saved with the model, rather
than creating an association for the new item immediately.