[C] Refactor error handling
This commit aims to improve exception handling in a number of contexts,
including:

* Exceptions thrown in the server-side render
* Exceptions thrown in the client-side application
* Exceptions returned by the Manifold API

Exceptions thrown in the browser are handled using a high-level React
error boundary in the Manifold container. The server-side render catches
exceptions, but instead of using react-redbox to render them, it
organizes them and renders them with the FatalError container. This
container has been expanded to render the different kinds of stack
traces that we need to see (js traces, component trees, and json stack
traces coming from the API). This commit also adjusts the complex
rendering chain in server-client to look for a FatalError in the store
after the entire component rendering tree is complete, and to
re-render a FatalError if it finds one. This means that components deep
in the tree can dispatch a 404 error and the SSR will return the correct
status code and message. Finally, this commit introduces a new spot
in the store for storing fatal errors. These were previously stored as
part of the Notification reducer.

Fixes #1370