[B] Clean related editor session on model  deletion
When a model is deleted, it is typically removed from the store by way
of an ENTITY_STORE_REMOVE action. This commit adjusts the entityEditor
reducer to listen for this action. If the entity being removed matches
any entity for which there is an open entityEditor session, that
session will be marked as "clean" regardless of whether or not it is
dirty. This change, in effect, prevents a dirty warning from appearing
after a model has been removed from the store, but before the UI
unmounts the edit form.

Fixes #813