[B] Address regression when navigating between notes
This commit moves annotation filtering into getDerivedStateFromProps
while continuing to check if there's an invalid annotation in the URL
after the component has updated. Prior to this change, the check was
happening before the annotations for a new text section were filtered
and stored in state, which cause the annotation hash to be stripped
every time.

Fixes #823