[B] TextNode annotation queries substring
Currently this looks for an exact `annotation-ids` match to
the id in the url hash.  This is an issue when a single TextNode
has multiple annotation ids on it.  Ultimately this results in a
null element being passed to `smoothScroll`, which has no protection
against a null element.
This commit adjusts the TextNode `doScroll` function to
query elements by `data-annotation-ids` substring presence and also
adds a guard to `smoothScroll` in any case where a null element is
passed to it

Fixes #1571