[B] Correct 404 handling for nested routes
This commit replaces the stock react-router switch with a custom switch
within the childRoutes method. The custom switch will throw a 404 if no
child routes match, and if the parent route does not match _exactly_.
This change addresses the fact that an inexact parent route with children
will not 404 in some cases. Eg, read/foo/section/bar is valid, and has
child routes (so is therefore not exact). Prior to this change,
read/foo/section/bar/some/garbage would not 404.