[C] Only wrap project import in transaction
With the latest changes to the dev loader, if something fails
at any point in the `#load` process the whole thing is rolled
back.  The rest of the seed data, besides projects, seems safe to
commit immediately.  This also enables being able to work with the
app while project data is loading.