[C] Prepare for Client reorganization
We temporarily render placeholders in routes to make reorg more
straight-forward. Some one-off imports are normalized in advance
of running a jscodeshift to adjust all component/container imports.
We remove the CLIENT_LOADED dispatch, which is causing problems
under this refactor, and we do a little work to reduce circular
depedencies in advance of the reorganization.