[B] Return placeholder when attachment processing
Previously this was always returning the original attachment version
for _all_ style keys that did not have a version associated with them.
This also included when attachments were in the queue for processing.

This commit adds new placeholder images and will return those URLs
for a given attachment style while an image is queued for processing.
Other attachment types will return nil immediately.
Fixes #1699