[C] Abstract .project-list-container into general section wrapper
This commit creates .entity-section-wrapper and nested classes to
manage a common layout across the frontend of a section title,
description, body, pagination, and filters. This change
abstracts the existing styles of project list and project collection
containers so that these styles may be used generally for any
similar high-level layout of page elements.

Components and containers are updated to reflect this change in
classNames and also to apply BEM naming conventions (e.g.
.entity-section-wrapper__heading), which helps avoid excessive scoping
of selectors and leave elements within things like the heading
unaffected.